requirements.txt
Django==5.2.8
pip==22.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1 
pyjwt==2.10.1
django-cors-headers==4.9.0